face_recognition_greeter_app/requirements.txt
pymediawiki
boto3
cmake
black
face_recognition
opencv-contrib-python
numpy
